HARP-12700: Extract loadImage out of ImageCache as free function.
